% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{astro_body_code}
\alias{astro_body_code}
\title{Get the integer code for a celestial body by name}
\usage{
astro_body_code(name)
}
\arguments{
\item{name}{One of the following strings: Sun, Moon, Mercury, Venus, Earth, Mars, Jupiter, Saturn, Uranus, Neptune, Pluto, EMB, SSB.}
}
\value{
If \code{name} is one of the listed strings (case-sensitive), the returned value is the corresponding integer code (see \link{astro_body}), otherwise it is \code{-1}.
}
\description{
Returns the integer code corresponding to the given English name.
}
\examples{
astro_body_code("Sun")
astro_body_code("Neptune")
}
