% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{astro_body_name}
\alias{astro_body_name}
\title{Get the name of a celestial body}
\usage{
astro_body_name(body)
}
\arguments{
\item{body}{The celestial body whose name is to be found.}
}
\value{
The English-language name of the celestial body, or \code{""} if the body is not valid.
}
\description{
Finds the name of a celestial body.
}
\examples{
astro_body_name(astro_body["SUN"])
astro_body_name(9L)
}
