% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{astro_current_time}
\alias{astro_current_time}
\title{Current UTC time according to Astronomy Engine}
\usage{
astro_current_time()
}
\value{
A \code{POSIXct} value representing the current UTC date and time.
}
\description{
Uses the computer's system clock to find the current UTC date and time.
Converts that date and time to a POSIXct value and returns the result.
Callers can pass this value to other Astronomy Engine functions to calculate
current observational conditions.
}
\details{
On supported platforms (Linux/Unix, Mac, Windows), the time is measured with
microsecond resolution.
}
\examples{
astro_current_time()
}
