% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{astro_identity_matrix}
\alias{astro_identity_matrix}
\title{Create an identity rotation matrix}
\usage{
astro_identity_matrix()
}
\value{
A list with components:
\describe{
\item{rot}{A 3x3 rotation matrix}
\item{status}{Status code (0 = success)}
}
}
\description{
Returns a rotation matrix that has no effect on orientation.
This matrix can be the starting point for other operations,
such as using a series of calls to \code{\link[=astro_pivot]{astro_pivot()}} to
create a custom rotation matrix.
}
\examples{
# Create an identity matrix
id <- astro_identity_matrix()
id
}
