% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{astro_rotation_ECL_EQJ}
\alias{astro_rotation_ECL_EQJ}
\title{Rotation Matrix from ECL to EQJ}
\usage{
astro_rotation_ECL_EQJ()
}
\value{
A rotation matrix that converts ECL to EQJ.
}
\description{
Calculates a rotation matrix from J2000 mean ecliptic (ECL) to J2000 mean
equator (EQJ).
}
\details{
This is one of the family of functions that returns a rotation matrix
for converting from one orientation to another.

\strong{Source}: ECL = ecliptic system, using equator at J2000 epoch.

\strong{Target}: EQJ = equatorial system, using equator at J2000 epoch.
}
\examples{
astro_rotation_ECL_EQJ()

}
