% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{astro_rotation_EQJ_GAL}
\alias{astro_rotation_EQJ_GAL}
\title{Rotation Matrix from EQJ to GAL}
\usage{
astro_rotation_EQJ_GAL()
}
\value{
A rotation matrix that converts EQJ to GAL.
}
\description{
Calculates a rotation matrix from J2000 mean equator (EQJ) to galactic (GAL).
}
\details{
This is one of the family of functions that returns a rotation matrix
for converting from one orientation to another.

\strong{Source}: EQJ = equatorial system, using the equator at the J2000 epoch.

\strong{Target}: GAL = galactic system (IAU 1958 definition).
}
\examples{
astro_rotation_EQJ_GAL()

}
