% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{astro_rotation_GAL_EQJ}
\alias{astro_rotation_GAL_EQJ}
\title{Rotation Matrix from GAL to EQJ}
\usage{
astro_rotation_GAL_EQJ()
}
\value{
A rotation matrix that converts GAL to EQJ.
}
\description{
Calculates a rotation matrix from galactic (GAL) to J2000 mean equator (EQJ).
}
\details{
This is one of the family of functions that returns a rotation matrix
for converting from one orientation to another.

\strong{Source}: GAL = galactic system (IAU 1958 definition).

\strong{Target}: EQJ = equatorial system, using the equator at the J2000 epoch.
}
\examples{
astro_rotation_GAL_EQJ()

}
