% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomyengine-package.R
\docType{package}
\name{astronomyengine-package}
\alias{astronomyengine}
\alias{astronomyengine-package}
\title{astronomyengine: R Bindings to the 'Astronomy Engine' C Library}
\description{
Provides access to the 'Astronomy Engine' C library (\url{https://github.com/cosinekitty/astronomy}) by Don Cross. The library calculates positions of the Sun, Moon, and planets, and predicts astronomical events such as rise/set times, lunar phases, equinoxes, solstices, eclipses, and transits. It is based on the 'VSOP87' planetary model and is accurate to within approximately one arcminute. This package bundles the single-file C source so that other R packages can link against it via 'LinkingTo' without shipping their own copy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mitchelloharawild/astronomyengine}
  \item Report bugs at \url{https://github.com/mitchelloharawild/astronomyengine/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

Authors:
\itemize{
  \item Don Cross (Author of the bundled Astronomy Engine C library) [copyright holder]
}

}
\keyword{internal}
